So much image resizing!
With these updates, I wanted to get my images to remain responsive, but appear to line up better on the homepage so they weren't all broken and different sizes. I was able to achieve this through applying different Bootstrap classes and one media query. Also, made a couple of edits to paragraphs to make the information as simple to understand as possible.